Дорабатывает отправку формы, сброс формы и карты
