+++ js/validate-form.js
@@ -1,9 +1,5 @@
-import {sendData} from './api.js';
-
 import {
   adForm,
-  mapFilters,
-  submitButtonForm
 } from './form.js';
 
 import {
@@ -11,7 +7,6 @@ import {
   VALUE_OPTION_NOT_FOR_GUESTS,
   REGISTRATION_HOURS
 } from './contants.js';
-import {isEscape} from './utils.js';
 
 const minPrice = {
   bungalow: 0,
@@ -137,65 +132,9 @@ const updateTimeInOut = (evt) => {
 
 formElementTime.addEventListener('change', updateTimeInOut);
 
-const popupOnSuccess = document.querySelector('#success').content
-  .querySelector('.success').cloneNode(true);
-const popupOnError = document.querySelector('#error').content
-  .querySelector('.error').cloneNode(true);
-
-function onPopupSuccessClick () {
-  popupOnSuccess.remove();
-}
-
-function onPopupSuccessEscDown (evt) {
-  if (isEscape(evt)) {
-    popupOnSuccess.remove();
-  }
-}
-
-function onPopupErrorEscClick () {
-  popupOnError.remove();
-}
-
-function onPopupErrorEscDown (evt) {
-  if (isEscape(evt)) {
-    popupOnError.remove();
-  }
-}
-
-const onSuccess = () => {
-  submitButtonForm.disabled = false;
-  submitButtonForm.textContent = 'Опубликовать';
-  document.body.insertAdjacentElement('beforeend', popupOnSuccess);
-  const popupOnSuccessElement = document.querySelector('.success');
-  document.addEventListener('keydown', onPopupSuccessEscDown);
-  popupOnSuccessElement.addEventListener('click', onPopupSuccessClick);
-
-  adForm.reset();
-  mapFilters.reset();
-};
-
-const onFail = () => {
-  submitButtonForm.disabled = false;
-  submitButtonForm.textContent = 'Опубликовать';
-  document.body.insertAdjacentElement('beforeend', popupOnError);
-  const popupOnErrorElement = document.querySelector('.error');
-  document.addEventListener('keydown', onPopupErrorEscDown);
-  popupOnErrorElement.addEventListener('click', onPopupErrorEscClick);
-};
-
-adForm.addEventListener('submit', (evt) => {
-  evt.preventDefault();
-  const isValid = pristine.validate();
-  if (isValid) {
-    submitButtonForm.disabled = true;
-    submitButtonForm.textContent = 'Публикуем ...';
-    const formData = new FormData(adForm);
-    sendData(onSuccess, onFail, formData);
-  }
-});
-
 export {
   price,
   kindType,
-  minPrice
+  minPrice,
+  pristine
 };
